move helper function to utils.py
